LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-first-textarea-002.tentative-expected.txt








FAIL .target > * 1 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 10
FAIL .target > * 2 assert_equals:
<textarea data-offset-x="10" class="inner"></textarea>
offsetLeft expected 10 but got 20
FAIL .target > * 3 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 15
FAIL .target > * 4 assert_equals:
<textarea data-offset-x="10" class="inner">X X X X X</textarea>
offsetLeft expected 10 but got 25
FAIL .target > * 5 assert_equals:
<span data-offset-x="60"></span>
offsetLeft expected 60 but got 15
FAIL .target > * 6 assert_equals:
<textarea data-offset-x="10" class="inner" placeholder="X X X X X"></textarea>
offsetLeft expected 10 but got 25
FAIL .target > * 7 assert_equals:
<span data-offset-x="25"></span>
offsetLeft expected 25 but got 15
FAIL .target > * 8 assert_equals:
<textarea data-offset-x="10" class="inner" style="font-size: 100px;"></textarea>
offsetLeft expected 10 but got 25
FAIL .target > * 9 assert_equals:
<span data-offset-x="25"></span>
offsetLeft expected 25 but got 10
FAIL .target > * 10 assert_equals:
<textarea data-offset-x="10" class="inner" style="font-size: 100px;">X X X X X</textarea>
offsetLeft expected 10 but got 20
FAIL .target > * 11 assert_equals:
<span data-offset-x="25"></span>
offsetLeft expected 25 but got 10
FAIL .target > * 12 assert_equals:
<textarea data-offset-x="10" class="inner" style="font-size: 100px;" placeholder="X X X X X"></textarea>
offsetLeft expected 10 but got 20
